pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sigmapy"
version = "0.1.0"
description = "Beginner-friendly tutorials and examples for Ergo blockchain development in Python"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Ergo Community", email = "community@ergoplatform.org"},
]
keywords = ["ergo", "blockchain", "tutorial", "examples", "sigma"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "requests>=2.25.0",
    "typing-extensions>=4.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.910",
    "pre-commit>=2.0",
]
docs = [
    "sphinx>=4.0",
    "sphinx-rtd-theme>=1.0",
    "myst-parser>=0.18",
]
full = [
    "ergo-lib-python>=0.9.0",
]

[project.urls]
Homepage = "https://github.com/ergoplatform/sigmapy"
Documentation = "https://github.com/ergoplatform/sigmapy"
Repository = "https://github.com/ergoplatform/sigmapy"
Issues = "https://github.com/ergoplatform/sigmapy/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true